+++ packages/core/src/utils/index.ts
@@ -1,11 +1,11 @@
 import Dragger from './Dragger';
-import Sorter from './Sorter';
+import Sorter from './sorter/Sorter';
 import Resizer from './Resizer';
 import * as mixins from './mixins';
 import { Module } from '../abstract';
 import EditorModel from '../editor/model/Editor';
-import ComponentSorter from './ComponentSorter';
-import StyleManagerSorter from './StyleManagerSorter';
+import ComponentSorter from './sorter/ComponentSorter';
+import StyleManagerSorter from './sorter/StyleManagerSorter';
 
 export default class UtilsModule extends Module {
   Sorter = Sorter;
